+++ test/unit/Timestamp.js
@@ -22,4 +22,11 @@ describe("Timestamp", function () {
         expect(timestamp.seconds.toInt()).to.be.eql(1);
         expect(timestamp.nanos.toInt()).to.be.eql(1);
     });
+
+    it("fromDate()", function () {
+        let timestamp = Timestamp.fromDate(999999998);
+
+        expect(timestamp.seconds.toInt()).to.be.eql(0);
+        expect(timestamp.nanos.toInt()).to.be.eql(999999998);
+    });
 });